third_party/blink/web_tests/editing/selection/caret-in-empty-inline-1-expected.txt

Bug 85793: Caret is not rendered in empty inline contenteditable elements

This test verifies that an empty inline contenteditable element gets a valid caret rect.



PASS caretRect.left is 8
PASS caretRect.top is 160
PASS caretRect.width is 1
PASS caretRect.height is 20
PASS successfullyParsed is true

TEST COMPLETE
